+++ async/src/test/kotlin/ktx/async/dispatchersTest.kt
@@ -5,6 +5,7 @@ import com.badlogic.gdx.utils.Timer
 import com.badlogic.gdx.utils.async.AsyncExecutor
 import com.nhaarman.mockitokotlin2.verify
 import io.kotlintest.mock.mock
+import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Runnable
 import kotlinx.coroutines.async
@@ -24,6 +25,7 @@ import java.util.concurrent.atomic.AtomicLong
 /**
  * Base class with coroutine dispatcher tests.
  */
+@OptIn(DelicateCoroutinesApi::class)
 abstract class CoroutineDispatcherTest : AsyncTest() {
   abstract val tested: KtxDispatcher
   abstract fun getExecutorThread(): Thread