+++ log/src/main/kotlin/ktx/log/log.kt
@@ -6,16 +6,23 @@ import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
 
+/** Debug logging tag. */
+const val DEBUG = "DEBUG"
+/** Info logging tag. */
+const val INFO = "INFO"
+/** Error logging tag. */
+const val ERROR = "ERROR"
+
 /**
  * Logs a message on the debug level.
- * @param tag will proceed the message. Defaults to "DEBUG" in square braces.
+ * @param tag will proceed the message. Defaults to "DEBUG".
  * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_DEBUG
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun debug(tag: String = "[DEBUG]", message: () -> String) {
+inline fun debug(tag: String = DEBUG, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(tag, message())
 }
@@ -23,28 +30,28 @@ inline fun debug(tag: String = "[DEBUG]", message: () -> String) {
 /**
  * Logs a message on the debug level.
  * @param cause its stack trace will be printed.
- * @param tag will proceed the message. Defaults to "DEBUG" in square braces.
+ * @param tag will proceed the message. Defaults to "DEBUG".
  * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_DEBUG
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun debug(cause: Throwable, tag: String = "[DEBUG]", message: () -> String) {
+inline fun debug(cause: Throwable, tag: String = DEBUG, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(tag, message(), cause)
 }
 
 /**
  * Logs a message on the info level.
- * @param tag will proceed the message. Defaults to "INFO" in square braces.
+ * @param tag will proceed the message. Defaults to "INFO".
  * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_INFO
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun info(tag: String = "[INFO] ", message: () -> String) {
+inline fun info(tag: String = INFO, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(tag, message())
 }
@@ -52,27 +59,28 @@ inline fun info(tag: String = "[INFO] ", message: () -> String) {
 /**
  * Logs a message on the info level.
  * @param cause its stack trace will be printed.
- * @param tag will proceed the message. Defaults to "INFO" in square braces.
+ * @param tag will proceed the message. Defaults to "INFO".
  * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_INFO
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun info(cause: Throwable, tag: String = "[INFO] ", message: () -> String) {
+inline fun info(cause: Throwable, tag: String = INFO, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(tag, message(), cause)
 }
 
 /**
  * Logs a message on the error level.
+ * @param tag will proceed the message. Defaults to "ERROR".
  * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_ERROR
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun error(tag: String = "[ERROR]", message: () -> String) {
+inline fun error(tag: String = ERROR, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(tag, message())
 }
@@ -80,49 +88,44 @@ inline fun error(tag: String = "[ERROR]", message: () -> String) {
 /**
  * Logs a message on the error level.
  * @param cause its stack trace will be printed.
- * @param tag will proceed the message. Defaults to "ERROR" in square braces.
+ * @param tag will proceed the message. Defaults to "ERROR".
  * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
  *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
  * @see Application.LOG_ERROR
  * @see Application.getLogLevel
  */
 @OptIn(ExperimentalContracts::class)
-inline fun error(cause: Throwable, tag: String = "[ERROR]", message: () -> String) {
+inline fun error(cause: Throwable, tag: String = ERROR, message: () -> String) {
   contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
   if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(tag, message(), cause)
 }
 
 /**
- * A simple logging utility class which caches its tags.
- * @param tag name of the logger included in log tags. Proceeded with a prefix of the specified log type.
- * @param debugPrefix will proceed [tag] in debug logs.
- * @param infoPrefix will proceed [tag] in info logs.
- * @param errorPrefix will proceed [tag] in error logs.
+ * A simple logging utility class that includes its [name] in the logged messages.
+ * @param name name of the logger prepended to each logged message.
+ * @param debugTag will be passed as the message tag on debug logs.
+ * @param infoTag will be passed as the message tag on info logs.
+ * @param errorTag will be passed as the message tag on error logs.
  */
 @Suppress("LeakingThis")
 open class Logger(
-  open val tag: String,
-  debugPrefix: String = "[DEBUG] ",
-  infoPrefix: String = "[INFO]  ",
-  errorPrefix: String = "[ERROR] "
+  open val name: String,
+  open val debugTag: String = DEBUG,
+  open val infoTag: String = INFO,
+  open val errorTag: String = ERROR
 ) {
-  // Implementation note: tags are not private as they are referenced by the inlined methods.
-  /**
-   * Will proceed all debug logs.
-   */
-  open val debugTag = "$debugPrefix$tag"
-
-  /**
-   * Will proceed all info logs.
-   */
-  open val infoTag = "$infoPrefix$tag"
-
-  /**
-   * Will proceed all error logs.
-   */
-  open val errorTag = "$errorPrefix$tag"
+  // Implementation note: tags and some internal methods are not private as they are referenced by the inlined methods
+  // and can be overridden.
 
   // TODO As of Kotlin 1.3, contracts are not allows in operators. Modify invoke methods to include contracts.
+
+  /**
+   * This method is called before passing the message to application logger. It can be overridden to modify the
+   * pattern of the logged message. By default, it will prepend the logger name and a dash to the original message.
+   * @param message original message that should be logged.
+   * @return formatted message ready for logging.
+   */
+  open fun buildMessage(message: String): String = "$name - $message"
 
   /**
    * Logs a message on the info level.
@@ -132,7 +135,7 @@ open class Logger(
    * @see Application.getLogLevel
    */
   inline operator fun invoke(message: () -> String) {
-    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, message())
+    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, buildMessage(message()))
   }
 
   /**
@@ -144,91 +147,89 @@ open class Logger(
    * @see Application.getLogLevel
    */
   inline operator fun invoke(cause: Throwable, message: () -> String) {
-    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, message(), cause)
+    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, buildMessage(message()), cause)
   }
-}
 
-// TODO In Kotlin 1.4, inlined Logger extension methods should be added directly to the Logger class.
+  /**
+   * Logs a message on the debug level.
+   * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_DEBUG
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun debug(message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(debugTag, buildMessage(message()))
+  }
 
-/**
- * Logs a message on the debug level.
- * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_DEBUG
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.debug(message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(debugTag, message())
-}
+  /**
+   * Logs a message on the debug level.
+   * @param cause its stack trace will be printed.
+   * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_DEBUG
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun debug(cause: Throwable, message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(debugTag, buildMessage(message()), cause)
+  }
 
-/**
- * Logs a message on the debug level.
- * @param cause its stack trace will be printed.
- * @param message inlined lambda which will be evaluated only if debug logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_DEBUG
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.debug(cause: Throwable, message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_DEBUG) Gdx.app.debug(debugTag, message(), cause)
-}
+  /**
+   * Logs a message on the info level.
+   * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_INFO
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun info(message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, buildMessage(message()))
+  }
 
-/**
- * Logs a message on the info level.
- * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_INFO
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.info(message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, message())
-}
+  /**
+   * Logs a message on the info level.
+   * @param cause its stack trace will be printed.
+   * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_INFO
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun info(cause: Throwable, message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, buildMessage(message()), cause)
+  }
 
-/**
- * Logs a message on the info level.
- * @param cause its stack trace will be printed.
- * @param message inlined lambda which will be evaluated only if info logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_INFO
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.info(cause: Throwable, message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_INFO) Gdx.app.log(infoTag, message(), cause)
-}
+  /**
+   * Logs a message on the error level.
+   * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_ERROR
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun error(message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(errorTag, buildMessage(message()))
+  }
 
-/**
- * Logs a message on the error level.
- * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_ERROR
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.error(message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(errorTag, message())
-}
-
-/**
- * Logs a message on the error level.
- * @param cause its stack trace will be printed.
- * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
- *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
- * @see Application.LOG_ERROR
- * @see Application.getLogLevel
- */
-@OptIn(ExperimentalContracts::class)
-inline fun Logger.error(cause: Throwable, message: () -> String) {
-  contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
-  if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(errorTag, message(), cause)
+  /**
+   * Logs a message on the error level.
+   * @param cause its stack trace will be printed.
+   * @param message inlined lambda which will be evaluated only if error logs are currently on. The string result of this
+   *    function will be created ONLY when needed, reducing the impact of creating new strings at runtime.
+   * @see Application.LOG_ERROR
+   * @see Application.getLogLevel
+   */
+  @OptIn(ExperimentalContracts::class)
+  inline fun error(cause: Throwable, message: () -> String) {
+    contract { callsInPlace(message, InvocationKind.AT_MOST_ONCE) }
+    if (Gdx.app.logLevel >= Application.LOG_ERROR) Gdx.app.error(errorTag, buildMessage(message()), cause)
+  }
 }
 
 /**