+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
+#### 1.10.0-b2
+
+- **[UPDATE]** Updated to Gradle 7.0.2.
+- **[UPDATE]** Updated to Kotlin 1.5.21.
+- **[UPDATE]** Updated to Kotlin Coroutines 1.5.1.
+- **[UPDATE]** Updated to VisUI 1.5.0.
+- **[CHANGE]** The JVM target compatibility of all modules was set to `1.8`, since Java `1.6` target was deprecated
+and is incompatible with the latest coroutines library. Note that source compatibility level is still set to `1.7`.
+- **[FEATURE]** (`ktx-assets`) Added `discard` lambda parameter to the `pool` factory method that is invoked after 
+an object is rejected from the created `Pool`.
+- **[FEATURE]** (`ktx-scene2d`) Added support for `ParticleEffectActor` in Scene2D DSL via `particleEffect` factory method.
+- **[FEATURE]** (`ktx-log`) Added `DEBUG`, `INFO` and `ERROR` constants with default logger tags.
+- **[CHANGE]** (`ktx-log`) Default tags are no longer wrapped in square brackets. This caused a change of the logs structure.
+- **[CHANGE]** (`ktx-log`) `Logger.tag` variable was renamed to `Logger.name`.
+- **[CHANGE]** (`ktx-log`) Log prefix variables from `Logger` were removed. Message tags are now configurable directly through the constructor.
+- **[CHANGE]** (`ktx-log`) Logger name is no longer a part of the message tag. Instead, it is now prepended to the message.
+This causes a change of the logs structure.
+- **[FEATURE]** (`ktx-log`) Added `Logger.buildMessage` method that allows to modify logged message structure. This method
+can be overridden when extending the `Logger` class, simplifying the usage of custom message formats.
+
 #### 1.10.0-b1
 
 - **[UPDATE]** Updated to LibGDX 1.10.0.